+++ akka-stream-tests/src/test/scala/akka/stream/scaladsl/SourceSpec.scala
@@ -50,41 +50,6 @@ class SourceSpec extends StreamSpec with DefaultTimeout {
     }
   }
 
-  "Future Source" must {
-    import akka.stream.impl.fusing.GraphStages.FutureFlattenSource
-
-    "flatten elements" in Utils.assertAllStagesStopped {
-      implicit def ec = materializer.executionContext
-
-      val subSource: Source[Int, String] =
-        Source(List(1, 2, 3)).mapMaterializedValue(_ ⇒ "foo")
-
-      val futureSource = new FutureFlattenSource(Future(subSource))
-      val source: Source[Int, Future[String]] = Source.fromGraph(futureSource)
-
-      val materialized = Promise[String]()
-      val watched: Source[Int, NotUsed] = source.watchTermination() { (m, d) ⇒
-        materialized.completeWith(d.flatMap(_ ⇒ m))
-        NotUsed
-      }
-
-      val p = watched.runWith(Sink.asPublisher(false))
-      val c = TestSubscriber.manualProbe[Int]()
-      p.subscribe(c)
-
-      val sub = c.expectSubscription()
-      sub.request(5)
-
-      c.expectNext(1)
-      c.expectNext(2)
-      c.expectNext(3)
-
-      c.expectComplete()
-
-      Await.result(materialized.future, 3.seconds) should ===("foo")
-    }
-  }
-
   "Empty Source" must {
     "complete immediately" in {
       val p = Source.empty.runWith(Sink.asPublisher(false))
